#3_UpdatePemeriksaan

+++ database/migrations/2023_05_27_121311_create_periksa_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->date('tglPeriksa');
             $table->unsignedBigInteger('mati');
             $table->unsignedBigInteger('culling');
-            $table->unsignedBigInteger('bobot');
+            $table->float('bobot',8,2);
             $table->unsignedBigInteger('pakan');
             $table->timestamps();
         });
